# docs/authors.swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bookstore - Authors API",
    "version": "1.0.0",
    "description": "Author management for the Bookstore platform.",
    "contact": {
      "name": "Bookstore Platform Team",
      "email": "satyasovan.tripathy@postman.com"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "https://bo-d225e84afdf94a038e1955bfff5e8c50.ecs.us-east-1.on.aws",
      "description": "AWS dev (ECS Express)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Authors",
      "description": "Author operations"
    }
  ],
  "paths": {
    "/api/authors": {
      "get": {
        "summary": "List all authors",
        "operationId": "listAuthors",
        "tags": [
          "Authors"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "A list of authors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new author",
        "operationId": "addAuthor",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/authors/{id}": {
      "get": {
        "summary": "Get an author by ID",
        "operationId": "getAuthorById",
        "tags": [
          "Authors"
        ],
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Author found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "404": {
            "description": "Author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an author",
        "operationId": "updateAuthor",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Author updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "404": {
            "description": "Author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an author",
        "operationId": "deleteAuthor",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Author deleted"
          },
          "404": {
            "description": "Author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Resource identifier (Mongo ObjectId)",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "books": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Book ObjectId reference"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "name"
        ]
      },
      "AuthorInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "name"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}